ARM OMAP3: rename arch_init_lowlevel to board_init_lowlevel
OMAP3 is the only architecture which has a arch_init_lowlevel in
which it invalidates the dcache. This can easily be done in
board_init_lowlevel aswell. Since on OMAP3 we are always executed
in SRAM we'll never need a board specific lowlevel_init. So the
easiest way of getting rid of this special handling is to just
rename the function.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>